requirements.txt
flask
pytest
pylint
python-dotenv
requests